test_model/requirements.txt
# ═══════════════════════════════════════════════════════════════
# requirements.txt
# Сегментация брюха тритона - RTX 3060 Ti
# ═══════════════════════════════════════════════════════════════

# Обработка изображений
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-image>=0.21.0

# Аугментация данных
albumentations>=1.3.0

# Визуализация и прогресс
matplotlib>=3.7.0
tqdm>=4.65.0

# Интерфейс приложения
gradio>=3.40.0

# Разметка данных
labelme>=5.4.0

# Примечание: PyTorch устанавливается отдельно с CUDA
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118